feat: separating functions and naming files by it's functions | fix: solving the null argument error
