+++ srcs/parsing/env_expansion.c
@@ -1,0 +1,6 @@
+#include "../../includes/minishell.h"
+/*
+A ideia é fazer a expansão dentro do function listener
+Que deve tratar os argumentos antes de passar para as funções
+*/
+
